Added: Ability to view employee payroll details in tabular format from local storage added

+++ EmployeePayrollApp/js/EmployeePayrollData.js
@@ -57,7 +57,7 @@ class EmployeePayrollData{
 
     toString() {
         const options = {
-            year: 'numeric', month: 'numeric', day: 'numeric'
+            year: 'numeric', month: 'long', day: 'numeric'
         };
         const empDate = !this.startDate ? "undefined" : this.startDate.toLocaleDateString("en-IN", options);
         return "name = " + this.name + ", profilePic = " + this.profilePic + ", salary = " + this.salary + ", gender = " + this.gender + ", department = " + this.department + ", startDate = " + empDate + ", notes= " + this.notes;
